Creating bridges
When a bridge is created on FE, the bridge DTO is pushed to RecordRepo.
It is sent to the "pushChanges" together with other records manually
from the Profile screen.  In the DB the data is stored in the tables
"bridges" and "bb_data_bridge_configurations".

Other DB tables will be used for Bat surveys, which is a work in
progress.