[BCGOVWILD-95] The "Bridge is for option" can't be unselected
Problem reason:
The onPress callback of a controlled BouncyCheckbox widget was using the
isChecker argument, which is explicitly prohibited by the docs:

    The onPress callback will NOT receive the next isChecked when
    disableBuiltInState is used.

Detected in the branch:
main

Conditions to reproduce the issue:
When creating a bridge, try clicking on a selected "Bridge is for"
option.  Nothing happens.

Possible workarounds without the fix:
No workarounds.

Changes:
Modify the onPress callback not to use arguments.

Risk factors:
Bridge Form screen.

Risk:
Low.